requirements.txt
asgiref==3.2.10
mysql-connector-python==8.0.24
Pillow==8.2.0
protobuf==3.16.0
pyserial==3.5
python-dotenv==0.17.1
pytz==2020.1
six==1.16.0
sqlparse==0.3.1
